Remove legacy rating system code
